// src/api_client/models/Account.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Diem Dev API Specification
 * Diem Dev API is REST API for client applications to interact the Diem blockchain.
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {exists, mapValues} from "../runtime";
/**
 * Core account resource, used for identifying account and transaction execution.
 * @export
 * @interface Account
 */
export interface Account {
  /**
   * Unsiged int64 type value
   * @type {string}
   * @memberof Account
   */
  sequenceNumber: string;
  /**
   * All bytes data are represented as hex-encoded string prefixed with `0x` and fulfilled with
   * two hex digits per byte.
   *
   * Different with `Address` type, hex-encoded bytes should not trim any zeros.
   * @type {string}
   * @memberof Account
   */
  authenticationKey: string;
}

export function AccountFromJSON(json: any): Account {
  return AccountFromJSONTyped(json, false);
}

export function AccountFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): Account {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    sequenceNumber: json["sequence_number"],
    authenticationKey: json["authentication_key"],
  };
}

export function AccountToJSON(value?: Account | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    sequence_number: value.sequenceNumber,
    authentication_key: value.authenticationKey,
  };
}