// src/api_client/models/MoveModule.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Diem Dev API Specification
 * Diem Dev API is REST API for client applications to interact the Diem blockchain. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    MoveModuleABI,
    MoveModuleABIFromJSON,
    MoveModuleABIFromJSONTyped,
    MoveModuleABIToJSON,
} from './MoveModuleABI';

/**
 * 
 * @export
 * @interface MoveModule
 */
export interface MoveModule {
    /**
     * All bytes data are represented as hex-encoded string prefixed with `0x` and fulfilled with
     * two hex digits per byte.
     * 
     * Different with `Address` type, hex-encoded bytes should not trim any zeros.
     * @type {string}
     * @memberof MoveModule
     */
    bytecode: string;
    /**
     * 
     * @type {MoveModuleABI}
     * @memberof MoveModule
     */
    abi?: MoveModuleABI;
}

export function MoveModuleFromJSON(json: any): MoveModule {
    return MoveModuleFromJSONTyped(json, false);
}

export function MoveModuleFromJSONTyped(json: any, ignoreDiscriminator: boolean): MoveModule {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'bytecode': json['bytecode'],
        'abi': !exists(json, 'abi') ? undefined : MoveModuleABIFromJSON(json['abi']),
    };
}

export function MoveModuleToJSON(value?: MoveModule | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'bytecode': value.bytecode,
        'abi': MoveModuleABIToJSON(value.abi),
    };
}